Fix stories of campaign tracking card and update hierarchy of stories
