+++ _dev/apps/ui/.storybook/preview.js
@@ -212,10 +212,18 @@ export const parameters = {
         'Google Ads Account',
         'Product Feed Page',
         'Campaign',
+        [
+          'Card',
+          'Tracking Card',
+          'Campaigns\' list',
+          'Form',
+          'Popins',
+        ],
         'Reporting',
         ['Key Metrics'],
         'Basic Components',
-        ['Whole application'],
+        'Whole application',
+        ['Overview'],
       ],
     },
   },
