+++ _dev/apps/ui/stories/smart-shopping-campaign-card.stories.ts
@@ -1,8 +1,4 @@
 import Campaign from '../src/components/campaigns/campaign-card.vue'
-import {campaigns} from '../.storybook/mock/campaigns-list';
-import {conversionTrackingIsTrue, conversionTrackingIsFalse} from '../.storybook/mock/campaigns';
-import cloneDeep from 'lodash.clonedeep';
-import {initialStateApp} from "../.storybook/mock/state-app";
 
 export default {
   title: 'Campaign/Card',
@@ -17,42 +13,8 @@ const Template = (args, { argTypes }) => ({
 });
 
 
-export const Enabled:any = Template.bind({});
-Enabled.args = {
+export const Card:any = Template.bind({});
+Card.args = {
   isEnabled: true,
   loading: false,
-};
-
-export const EnabledWithRemarketingSwitchOff:any = Template.bind({});
-EnabledWithRemarketingSwitchOff.args = {
-  isEnabled: true,
-  loading: false,
-  beforeMount(this: any) {
-    this.$store.state.app = cloneDeep(initialStateApp);
-    this.$store.state.campaigns = cloneDeep(conversionTrackingIsFalse);
-    this.$store.state.campaigns.campaigns = cloneDeep(campaigns);
-  },
-};
-
-export const EnabledWithRemarketingSwitchOn:any = Template.bind({});
-EnabledWithRemarketingSwitchOn.args = {
-  isEnabled: true,
-  loading: false,
-  beforeMount(this: any) {
-    this.$store.state.app = cloneDeep(initialStateApp);
-    this.$store.state.campaigns = cloneDeep(conversionTrackingIsTrue);
-    this.$store.state.campaigns.campaigns = cloneDeep(campaigns);
-    this.$store.state.campaigns.conversionActions = [{}];
-  },
-};
-
-export const AlertAboutMissingConversionActions:any = Template.bind({});
-AlertAboutMissingConversionActions.args = {
-  isEnabled: true,
-  loading: false,
-  beforeMount(this: any) {
-    this.$store.state.campaigns = cloneDeep(conversionTrackingIsTrue);
-    this.$store.state.campaigns.campaigns = cloneDeep(campaigns);
-    this.$store.state.campaigns.conversionActions = [];
-  },
 };
